requirements_dev.txt
-r requirements.txt

codecov==2.1.10
coloredlogs==14.0
coverage==5.3.1
django-naomi==0.8
flake8-blind-except==0.1.1
flake8-commas==2.0.0
flake8-docstrings==1.5.0
flake8-isort==4.0.0
flake8-pep3101==1.3.0
flake8-quotes==3.2.0
flake8-string-format==0.3.0
flake8==3.8.4
ipython==7.16.3
isort==5.7.0
mock==4.0.3
parameterized==0.8.1
pydocstyle==5.1.1
python-coveralls==2.9.3
sphinx==3.3.0
autopep8==1.5.4